# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Church ICT Service Management Platform API",
    "description": "RESTful API for managing ICT support tickets, equipment/room bookings, users, FAQs, announcements, and notifications.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Relative to deployment host" }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user account",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["name", "email", "password"],
            "properties": {
              "name": { "type": "string", "example": "David Developer" },
              "email": { "type": "string", "format": "email", "example": "david@church.org" },
              "password": { "type": "string", "format": "password", "example": "securePassword123" }
            }
          } } }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Invalid input" },
          "409": { "description": "Email already registered" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in and receive a JWT",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["email", "password"],
            "properties": {
              "email": { "type": "string", "format": "email" },
              "password": { "type": "string", "format": "password" }
            }
          } } }
        },
        "responses": {
          "200": { "description": "Authenticated" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get the logged-in user's profile",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Current user" }, "401": { "description": "Not authenticated" } }
      }
    },
...
    "/bookings": {
      "post": {
        "tags": ["Bookings"],
        "summary": "Request an equipment or room booking",
        "description": "Rejects the request with 409 if the item already has a Pending or Approved booking overlapping the requested window.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["item_id", "start_time", "end_time"],
            "properties": {
              "item_id": { "type": "integer" },
              "start_time": { "type": "string", "format": "date-time" },
              "end_time": { "type": "string", "format": "date-time" },
              "purpose": { "type": "string" }
            }
          } } }
        },
        "responses": {
          "201": { "description": "Booking requested" },
          "409": { "description": "Double-booking conflict" }
        }
      },
      "get": {
        "tags": ["Bookings"],
        "summary": "List bookings (own for Users; all for Staff/Admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "item_id", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "List of bookings" } }
      }
    },
    "/bookings/{id}/approve": {
      "patch": {
        "tags": ["Bookings"],
        "summary": "Approve a booking (Staff/Admin only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Approved" }, "409": { "description": "Conflicts with an approved booking" } }
      }
    },
    "/bookings/{id}/reject": {
      "patch": {
        "tags": ["Bookings"],
        "summary": "Reject a booking (Staff/Admin only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Rejected" } }
      }
    },
    "/bookings/{id}/cancel": {
      "patch": {
        "tags": ["Bookings"],
        "summary": "Cancel your own booking",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Cancelled" } }
      }
    },
...
    "/faqs": {
      "get": {
        "tags": ["FAQs"],
        "summary": "List FAQs (public)",
        "responses": { "200": { "description": "List of FAQs" } }
      },
      "post": {
        "tags": ["FAQs"],
        "summary": "Add an FAQ (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["question", "answer"],
            "properties": { "question": { "type": "string" }, "answer": { "type": "string" } }
          } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/faqs/{id}": {
      "delete": {
        "tags": ["FAQs"],
        "summary": "Delete an FAQ (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/announcements": {
      "get": {
        "tags": ["Announcements"],
        "summary": "List announcements (public)",
        "responses": { "200": { "description": "List of announcements" } }
      },
      "post": {
        "tags": ["Announcements"],
        "summary": "Publish an announcement (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["title", "body"],
            "properties": { "title": { "type": "string" }, "body": { "type": "string" } }
          } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/announcements/{id}": {
      "delete": {
        "tags": ["Announcements"],
        "summary": "Delete an announcement (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all users (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "List of users" } }
      }
    },
    "/admin/users/{id}/role": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Change a user's role (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["role"],
            "properties": { "role": { "type": "string", "enum": ["User", "Staff", "Admin"] } }
          } } }
        },
        "responses": { "200": { "description": "Updated" } }
      }
    },
    "/admin/users/{id}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a user (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/admin/stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Basic platform statistics (Admin only)",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Stats object" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}